pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "folium>=0.19.2",
    "google-auth>=2.37.0",
    "gspread>=6.1.4",
    "joblib>=1.4.2",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "scikit-learn>=1.6.0",
    "streamlit-folium>=0.24.0",
    "streamlit>=1.41.1",
    "trafilatura>=2.0.0",
    "numpy>=2.2.1",
    "python-dotenv>=1.0.1",
    "pywebpush>=2.0.3",
]
